vr首页数据加载、列表页代码编写

+++ app/src/main/java/com/expo/contract/VRDetailContract.java
@@ -11,12 +11,9 @@ public interface VRDetailContract {
             super(view);
         }
 
-        public abstract void setVrInfo(int id);
+        public abstract void setVrInfo(long id);
 
         public abstract void setVrResViews(int id);
-
-
-
     }
 
     interface View extends IView {
